# cmd/server/server.go
package server

import (
	"database/sql"

	"github.com/cloud-tech-develop/aura-back/modules/admin/enterprise"
	thirdparties "github.com/cloud-tech-develop/aura-back/modules/admin/third-parties"
	"github.com/cloud-tech-develop/aura-back/modules/admin/users"
	"github.com/cloud-tech-develop/aura-back/modules/catalog/brands"
	"github.com/cloud-tech-develop/aura-back/modules/catalog/categories"
	catalogproducts "github.com/cloud-tech-develop/aura-back/modules/catalog/products"
	"github.com/cloud-tech-develop/aura-back/shared/response"
	"github.com/cloud-tech-develop/aura-back/tenant"
	"github.com/gin-gonic/gin"
)

// Server holds the router and shared dependencies.
type Server struct {
	router    *gin.Engine
	db        *sql.DB
	tenantMgr *tenant.Manager
}

func corsMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Accept, Authorization")
		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}
		c.Next()
	}
}

// NewServer creates and configures the Gin router with global middleware.
func NewServer(db *sql.DB, tenantMgr *tenant.Manager) *Server {
	r := gin.Default()
	r.Use(corsMiddleware())

	return &Server{
		router:    r,
		db:        db,
		tenantMgr: tenantMgr,
	}
}

// RegisterModules mounts all module routes onto the router.
// All modules receive:
//   - a public group (no auth)
//   - a protected group (behind AuthMiddleware)
//
// To add a new module, call its Register() here.
func (s *Server) RegisterModules(
	enterpriseH *enterprise.Handler,
	userH *users.Handler,
	categoryH *categories.Handler,
	brandH *brands.Handler,
	productH *catalogproducts.Handler,
	thirdPartiesH *thirdparties.Handler,
) {
	// ── Health Check ──────────────────────────────────────────────────────────
	s.router.GET("/", func(c *gin.Context) {
		c.JSON(200, gin.H{"message": "hello world", "status": "ok"})
	})

	// ── Auth ─────────────────────────────────────────────────────────────────
	s.router.POST("/login", tenant.Login(s.db))

	// ── Static Files ──────────────────────────────────────────────────────────
	s.router.Static("/static", "./static")
	s.router.GET("/download/offline-pos", func(c *gin.Context) {
		c.FileAttachment("./static/bin/aura-pos-offline.exe", "aura-pos-offline.exe")
	})

	// ── Public routes (no middleware) ─────────────────────────────────────────
	public := s.router.Group("/")

	// ── Protected routes (JWT required) ───────────────────────────────────────
	protected := s.router.Group("/")
	protected.Use(tenant.AuthMiddleware())

	// ── Well-known protected routes ───────────────────────────────────────────
	protected.GET("/me", s.meHandler)

	// ── Feature Modules ───────────────────────────────────────────────────────
	enterprise.Register(public, protected, enterpriseH)
	users.Register(public, protected, userH)
	categories.Register(public, protected, categoryH)
	brands.Register(public, protected, brandH)
	catalogproducts.Register(public, protected, productH)
	thirdparties.Register(public, protected, thirdPartiesH)
}

// Run starts the HTTP server on the given address.
func (s *Server) Run(addr string) error {
	return s.router.Run(addr)
}

// meHandler returns the authenticated user's context info.
func (s *Server) meHandler(c *gin.Context) {
	response.OK(c, gin.H{
		"user_id":       c.GetInt64("user_id"),
		"enterprise_id": c.GetInt64("enterprise_id"),
		"slug":          c.GetString(string(tenant.TenantKey)),
		"email":         c.GetString("email"),
		"roles":         c.GetStringSlice("roles"),
	})
}