# infrastructure/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Aura Back API
  description: API documentation for Aura Back multi-tenant POS application
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/anomalyco/opencode/issues
servers:
  - url: http://localhost:8081
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
...
paths:
  /login:
    post:
      summary: Authenticate user and get JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Bad Request
        "401":
          description: Invalid credentials
  /me:
    get:
      summary: Get current authenticated user info
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    format: int64
                  enterprise_id:
                    type: integer
                    format: int64
                  slug:
                    type: string
                  email:
                    type: string
                  roles:
                    type: array
                    items:
                      type: string
        "401":
          description: Unauthorized

  # ============================================
  # ENTERPRISE MODULE (Public)
  # ============================================
...